Extract DigestHelper from ConfigHelper
