+++ server/app/helpers/config_helper.rb
@@ -1,18 +1,7 @@
-require 'symmetric-encryption'
-require 'bcrypt'
-
 module ConfigHelper
 
   def self.included(base)
     base.send :extend, ClassMethods
-  end
-
-  def salt
-    self.class.salt
-  end
-
-  def digest(string)
-    self.class.digest(string)
   end
 
   def config
@@ -20,25 +9,11 @@ module ConfigHelper
   end
 
   module ClassMethods
-
-    def salt
-      return @salt unless @salt.nil?
-      encrypted_salt = config[:salt]
-      if encrypted_salt
-        @salt = SymmetricEncryption.decrypt(encrypted_salt)
-      else
-        @salt = BCrypt::Engine.generate_salt
-        config[:salt] = SymmetricEncryption.encrypt(@salt)
-      end
-      @salt
-    end
-
-    def digest(string)
-      BCrypt::Engine.hash_secret(string, salt)
-    end
-
     def config
-      @config ||= Configuration
+      return @config if @config
+      @config = Configuration
+      @config.seed(Server.root.join('config/seed.yml'))
+      @config
     end
   end
 end
